Update SDK for iMXRT-10xx Family to SDK version

+++ middleware/eiq/deepviewrt/include/deepview_rpc.h
@@ -100,7 +100,7 @@ struct rpc_request {
      * physical address of the video memory for the current frame.  For shared
      * memory this is the offset into the shared memory object.
      */
-    uint32_t memory;
+    uint64_t memory;
 };
 
 /**
